Eliminate init function
With previous versons of React Native and Om,
it was necessary to set up a looping call to
requestAnimationFrame in order to get the iOS
UI to update automatically upon state or code
changes. (You'd have to click in the UI or
otherwise interact with it.) Now, this is
evidently no longer necessary.